// src/app/model/property.model.ts
export interface Property {
  id: string;
  title: string;
  property_image: string;
  property_image_thumb: string;
  short_description: string;
  long_description: string;
  main_amount: string;
  investment_amount: string;
  property_tenure: string;
  start_date: string;
  end_date: string;
  status: string;
  created_at: string;
  background_image: any;
  sno: number;
}